Fixed bugs about all the controller

+++ src/main/java/com/clinic/models/Receipt.java
@@ -10,7 +10,7 @@ public class Receipt {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long receipt_id;
+    private int receipt_id;
     private String receipt_description;
     private int patient_id;
     private int doctor_id;
@@ -18,16 +18,16 @@ public class Receipt {
     public Receipt() {
     }
 
-    public Receipt(Long id, String receipt_description) {
+    public Receipt(int id, String receipt_description) {
         this.receipt_id = id;
         this.receipt_description = receipt_description;
     }
 
-    public Long getReceipt_id() {
+    public int getReceipt_id() {
         return receipt_id;
     }
 
-    public void setReceipt_id(Long receipt_id) {
+    public void setReceipt_id(int receipt_id) {
         this.receipt_id = receipt_id;
     }
 
@@ -45,11 +45,11 @@ public class Receipt {
         return new Doctor().getDoctor_id();
     }
 
-    public Long getId() {
+    public int getId() {
         return receipt_id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.receipt_id = id;
     }
